+++ src/System.Security.Permissions/src/System.Security.Permissions.csproj
@@ -30,11 +30,6 @@
     <Compile Include="System\Diagnostics\PerformanceCounterPermissionAttribute.cs" />
     <Compile Include="System\Diagnostics\PerformanceCounterPermissionEntry.cs" />
     <Compile Include="System\Diagnostics\PerformanceCounterPermissionEntryCollection.cs" />
-    <Compile Include="System\DirectoryServices\DirectoryServicesPermission.cs" />
-    <Compile Include="System\DirectoryServices\DirectoryServicesPermissionAccess.cs" />
-    <Compile Include="System\DirectoryServices\DirectoryServicesPermissionAttribute.cs" />
-    <Compile Include="System\DirectoryServices\DirectoryServicesPermissionEntry.cs" />
-    <Compile Include="System\DirectoryServices\DirectoryServicesPermissionEntryCollection.cs" />
     <Compile Include="System\Drawing\Printing\PrintingPermission.cs" />
     <Compile Include="System\Drawing\Printing\PrintingPermissionAttribute.cs" />
     <Compile Include="System\Drawing\Printing\PrintingPermissionLevel.cs" />
@@ -192,7 +187,6 @@
     <Reference Include="System.Core" />
     <Reference Include="System.Data" />
     <Reference Include="System.Data.OracleClient" />
-    <Reference Include="System.DirectoryServices" />
     <Reference Include="System.Drawing" />
     <Reference Include="System.Net" />
     <Reference Include="System.Security" />
